% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Function performing the bootstrap}
\usage{
bootstrap(
  B,
  WB = FALSE,
  fu = function(x) x,
  type = "gumbel",
  B_fit = FALSE,
  alpha = NULL,
  x,
  fitted,
  residuals,
  start,
  loss,
  d = deriv_loss(loss),
  tuning,
  maxi,
  tau,
  lambda,
  K,
  ncores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{B_fit}{logical; if \code{FALSE}, the estimated coefficients for each bootstrap sample are returned in \code{ciraw}; if \code{TRUE}, the fitted values per sample are returned}

\item{alpha}{confidence level for bootstrap confidence intervals; if \code{alpha=NULL} the confidence level is set to 0.05}

\item{x}{for \code{\link{m}}: design matrix; for \code{\link{m_gam}}: NULL}

\item{fitted}{fitted values}

\item{residuals}{residuals}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{d}{derivative of chosen loss function divided by its argument}

\item{tuning}{set of tuning parameters}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{for \code{\link{m}}: initial lambda; for \code{\link{m_gam}}: smoothing parameter estimation method
as required by \code{\link[mgcv]{gam}}}

\item{K}{for \code{\link{m}}: quadratic penalty matrix; for \code{\link{m_gam}}: preceeding gam model with \code{fit=FALSE}}

\item{ncores}{number of cores to be used for parallelization}

\item{...}{further arguments which can be passed to \code{\link[mgcv]{gam}}}
}
\value{
\describe{
\item{\code{ciquantiles}}{bootstrap confidence intervals}
\item{\code{ciraw}}{estimated coefficients or fitted values for each bootstrap sample}
}
}
\description{
This function performs the bootstrap for the bootstrap samples generated via \code{\link{bootstrap_samples}}.
It can be called by \code{\link{m}} and by \code{\link{m_gam}} and allows for unpenalized as well as penalized models.
If \code{lambda} is chosen by cross-validation, the optimal value is estimated for each bootstrap sample.
}
