% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_confidence_interval.R
\name{bootstrap_confidence_interval}
\alias{bootstrap_confidence_interval}
\title{Function calculating the bootstrap confidence intervals}
\usage{
bootstrap_confidence_interval(
  n,
  alpha = NULL,
  est_coefs,
  bs_coefs,
  bs_ci_method = c("rev_percentile", "bca", "et_percentile-t", "sym_percentile-t"),
  B_nested = B,
  WB = FALSE,
  fu = function(x) x,
  type = "gumbel",
  x = NULL,
  y = NULL,
  start = NULL,
  loss = "nils",
  tuning = NULL,
  maxi = 5500,
  tau = 10^(-5),
  lambda = 0,
  K = NULL
)
}
\arguments{
\item{n}{number of observations}

\item{alpha}{confidence level for bootstrap confidence intervals; if \code{alpha=NULL} the confidence level is set to 0.05}

\item{est_coefs}{estimated coefficients}

\item{bs_coefs}{estimated coefficients for each bootstrap sample}

\item{bs_ci_method}{bootstrap confidence interval construction method:
\code{"rev_percentile"} = bootstrap reverse percentile, \code{"et_percentile-t"} = equal-tailed bootstrap percentile-t,
\code{"sym_percentile-t"} = symmetric bootstrap percentile-t}

\item{B_nested}{for \code{bs_ci_method = "et_percentile-t"}: number of nested bootstrap iterations per original bootstrap iteration}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{x}{design matrix}

\item{y}{response vector}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{tuning}{set of tuning parameters, if \code{NULL}, the tuning parameters are not set manually}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{vector of smoothing parameters; if \code{length(lambda>1)} they will be compared via cross-validation;
if \code{length(lambda)==2}, lambda is cross-validated on a log-scale between those values;
if \code{length(lambda>2)}, the values in \code{lambda} are cross-validated}

\item{K}{quadratic penalty matrix}
}
\value{
\describe{
\item{\code{bootstrap_ci}}{matrix containing the bootstrap confidence intervals for the different coefficients as rows}
}
}
\description{
This function calculates the bootstrap confidence intervals. Different construction methods can be chosen:
bootstrap reverse percentile, equal-tailed bootstrap percentile-t and symmetric bootstrap percentile-t.
The percentile confidence interval is already calculated in the function \code{\link{bootstrap}}.
}
