% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_samples.R
\name{bootstrap_samples}
\alias{bootstrap_samples}
\title{Function for generating the bootstrap samples}
\usage{
bootstrap_samples(
  B,
  WB = FALSE,
  fu = function(x) x,
  type = "gumbel",
  x,
  fitted,
  residuals
)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{x}{design matrix}

\item{fitted}{fitted values}

\item{residuals}{residuals}
}
\value{
\describe{
\item{\code{samples}}{bootstrap samples}
}
}
\description{
This function generates the bootstrap samples via residual bootstrap.
If \code{WB = TRUE}, wild bootstrap instead of residual bootstrap is performed.
}
