% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_coefs_distributions.R
\name{bs_coefs_distributions}
\alias{bs_coefs_distributions}
\title{Function plotting histograms, kdes and bootstrap confidence intervals for given bootstrapped coefficients}
\usage{
bs_coefs_distributions(
  nrep_plot = NULL,
  results_path = "./results",
  results_filename,
  plots_path = "./plots",
  plots_filename = results_filename
)
}
\arguments{
\item{nrep_plot}{number of repetitions to be plotted}

\item{results_path}{path to the folder which contains the results of the preceeding simulation study}

\item{results_filename}{filename of the result file}

\item{plots_path}{path to the folder where the plot should be saved}

\item{plots_filename}{filename of the resulting plot}
}
\description{
This function plots the histograms, kdes and bootstrap confidence intervals for given bootstrapped coefficients.
The true coefficients are indicated by blue vertical dashed lines. As result, a table containing the plots
for the different coefficients is saved.
}
