% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions.R
\name{compare_distributions}
\alias{compare_distributions}
\title{Function comparing the kdes of the bootstrapped residuals and the true density for the different losses}
\usage{
compare_distributions(
  n_plot = NULL,
  true_density,
  true_quantile,
  support,
  data_path = "./data",
  data_filename,
  results_path = "./results",
  results_filename,
  plots_path = "./plots",
  plots_filename = results_filename
)
}
\arguments{
\item{n_plot}{number of observations to plot}

\item{true_density}{true density function}

\item{true_quantile}{true quantile function}

\item{support}{support used for calculating the Kullback-Leibler divergences}

\item{data_path}{path to the folder which contains the data used for the preceeding simulation studies}

\item{data_filename}{filename of the data used for the preceeding simulation studies}

\item{results_path}{path to the folder which contains the results of the preceeding simulation studies
for the four different losses}

\item{results_filename}{filename (without the "_loss" ending) of the results files}

\item{plots_path}{path to the folder where the plot should be saved}

\item{plots_filename}{filename of the resulting plot}
}
\description{
This function plots the kdes of the bootstrapped residuals and the true density for the different losses.
Furthermore, it calculates the Kullback-Leibler divergences and plots QQ plots.
As result, a table of 4x2 plots is saved.
}
