% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_distributions_heteroskedastic.R
\name{compare_distributions_heteroskedastic}
\alias{compare_distributions_heteroskedastic}
\title{Function comparing the kdes of the bootstrapped residuals and the true densities for the different losses
for the heteroskedasticity cases in the Kemp study}
\usage{
compare_distributions_heteroskedastic(
  n_plot = NULL,
  bins = bins,
  data_path = "./data",
  data_filename,
  results_path = "./results",
  results_filename,
  plots_path = "./plots",
  plots_filename = results_filename
)
}
\arguments{
\item{n_plot}{number of observations to plot}

\item{bins}{parameter bins for \code{\link[ggplot2]{geom_histogram}}}

\item{data_path}{path to the folder which contains the data used for the preceeding simulation studies}

\item{data_filename}{filename of the data used for the preceeding simulation studies}

\item{results_path}{path to the folder which contains the results of the preceeding simulation studies
for the four different losses}

\item{results_filename}{filename (without the "_loss" ending) of the results files}

\item{plots_path}{path to the folder where the plot should be saved}

\item{plots_filename}{filename of the resulting plot}
}
\description{
This function plots the kdes of the bootstrapped residuals and the true densities for the different losses
for the heteroskedasticity cases in the Kemp study. For those cases, the true density depends on the regressor values.
Hence, there is for every observation another true density which needs to be calculated inside this function.
As result, a table of 4 plots is saved.
}
