% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Function calculating (asymptotic) confidence intervals}
\usage{
confidence_interval(
  x,
  obj,
  loss,
  alpha = 0.05,
  exact = FALSE,
  kernel = "gaussian"
)
}
\arguments{
\item{x}{design matrix}

\item{obj}{object generated by a previous call of \code{\link{m}} or \code{\link{m_gam}}}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{alpha}{confidence level}

\item{exact}{logical; for \code{loss="mean"}: \code{exact="TRUE"} can be chosen to calculate the exact
confidence interval under the normal distribution assumption}

\item{kernel}{for \code{loss="med"}: kernel for estimating the asymptotic covariance matrix (\code{gaussian} or \code{uniform})}
}
\value{
\describe{
\item{\code{ci}}{matrix containing the confidence intervals for the different coefficients as rows}
\item{\code{length}}{vector containing the length of the confidence intervals}
}
}
\description{
This functions calculates asymptotic (1-alpha) confidence intervals for the coefficients in different regression settings.
The chosen loss function determines the regression setting and for estimating the asymptotic covariance matrix
the results of previous calls of \code{\link{m}} / \code{\link{m_gam}} are used.
For \code{loss="mean"} it is also possible to calculate the exact (1-alpha) confidence interval under the normal distribution assumption.
Those confidence intervals need to be carefully interpreted since the asymptotic justifications demand for different assumptions
which do not have to be fullfilled.
}
