% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_rates.R
\name{coverage_rates}
\alias{coverage_rates}
\title{Function calculating the coverage rates of the asymptotic and bootstrap confidence intervals}
\usage{
coverage_rates(coefs, ci, bs_ci)
}
\arguments{
\item{coefs}{vector of true coefficients}

\item{ci}{list of matrices containing the asymptotic confidence intervals as rows for the different repetitions}

\item{bs_ci}{list of matrices containing the bootstrap confidence intervals as rows for the different repetitions}
}
\value{
\describe{
\item{\code{rates}}{matrix containing the coverage rates of the asymptotic confidence intervals and
of the bootstrap confidence intervals as columns}
}
}
\description{
This function calculates the coverage rates of the asymptotic and bootstrap confidence intervals.
}
