% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_loss.R
\name{deriv_loss}
\alias{deriv_loss}
\title{Function needed for calculating the estimates in the IRLS algorithm}
\usage{
deriv_loss(loss)
}
\arguments{
\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}
}
\value{
\describe{
\item{\code{d}}{derivative of chosen loss function divided by its argument as function}
}
}
\description{
The function calculates the derivative of the chosen loss function divided by its argument, constant factors are omitted.
This is needed to calculate the estimates of the coefficients in every iteration of the IRLS algorithm.
}
