% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cov.R
\name{est_cov}
\alias{est_cov}
\title{Function for estimating the asymptotic covariance matrix}
\usage{
est_cov(x, residuals, n, loss, tuning, j, correction = FALSE)
}
\arguments{
\item{x}{design matrix}

\item{residuals}{residuals}

\item{n}{number of observations}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012)}

\item{tuning}{set of tuning parameters}

\item{j}{internal index for tuning the IRLS algorithm}

\item{correction}{logical; if TRUE, the finite sample corrected residuals are returned}
}
\value{
\describe{
\item{\code{Omegahat}}{estimated asymptotic covariance matrix multiplied by \code{k^3/n}}
\item{\code{k}}{for \code{loss="nils"}: tuning parameter in last IRLS iteration, for \code{loss="kemp"}: inverse bandwith in last IRLS iteration}
}
}
\description{
This function calculates the asymptotic covariance matrix but returns the matrix
multiplied with the factor (k_n^3/n) to simplify the calculation of the asymptotic confidence intervals.
}
