% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Function performing the (penalized) mode, mean and median regression with lambda chosen outside the IRLS algorithm}
\usage{
m(
  x,
  y,
  start = NULL,
  loss = "nils",
  tuning = NULL,
  maxi = 5500,
  tau = 10^(-5),
  lambda = 0,
  K = NULL,
  B = 0,
  WB = FALSE,
  fu = function(x) x,
  type = "gumbel",
  B_fit = FALSE,
  alpha = NULL,
  cov = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{tuning}{set of tuning parameters, if \code{NULL}, the tuning parameters are not set manually}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{vector of smoothing parameters; if \code{length(lambda>1)} they will be compared via cross-validation;
if \code{length(lambda)==2}, lambda is cross-validated on a log-scale between those values;
if \code{length(lambda>2)}, the values in \code{lambda} are cross-validated}

\item{K}{quadratic penalty matrix}

\item{B}{number of bootstrap samples; if \code{B=0} no bootstrap will be performed}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{B_fit}{logical; if \code{FALSE}, the estimated coefficients for each bootstrap sample are returned in \code{bs_ciraw}; if \code{TRUE}, the fitted values per sample are returned}

\item{alpha}{confidence level for bootstrap confidence intervals; if \code{alpha=NULL} the confidence level is set to 0.05}

\item{cov}{logical; if \code{TRUE} and the type of loss function is either \code{"nils"} or \code{"kemp"},
the estimated asymptotic covariance matrix multiplied by k^3/n will be calculated}

\item{ncores}{number of cores to be used for parallelization}
}
\value{
\describe{
\item{\code{coefficients}}{estimated coefficients}
\item{\code{residuals}}{residuals}
\item{\code{fitted}}{fitted values}
\item{\code{weights}}{weights of final iteration in IRLS algorithm}
\item{\code{iter}}{number of iterations in IRLS algorithm}
\item{\code{converged}}{logical; if \code{TRUE}, the IRLS algorithm converged}
\item{\code{tuning}}{set of tuning parameters}
\item{\code{K}}{penalty matrix}
\item{\code{lambda}}{penalization parameter used for fitting (might be determined via cross-validation)}
\item{\code{cv}}{list containing the GCV scores and the estimated coefficients for the different cross-validated smoothing parameter values}
\item{\code{bs_ci}}{bootstrap confidence intervals}
\item{\code{bs_ciraw}}{estimated coefficients or fitted values for each bootstrap sample}
\item{\code{Omegahat}}{estimated asymptotic covariance matrix multiplied by \code{k^3/n}}
\item{\code{k}}{for \code{loss="nils"}: tuning parameter in last IRLS iteration, for \code{loss="kemp"}: inverse bandwith in last IRLS iteration}
}
}
\description{
This function performs the (possibly penalized) mode, median and mean regression where the penalization parameter lambda
can be a given value or can be chosen outside the IRLS algorithm via cross-validation.
}
