% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_control.R
\name{m_control}
\alias{m_control}
\title{Function for defining the tuning parameters and the starting values}
\usage{
m_control(x, y, loss, tuning, maxi, start, lambda, K)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{tuning}{set of tuning parameters, if \code{NULL}, the tuning parameters are not set manually}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{lambda}{penalty parameter}

\item{K}{quadratic penalty matrix}
}
\value{
\describe{
\item{\code{tuning}}{set of tuning parameters}
\item{\code{maxi}}{maximal number of iterations in the IRLS algorithm}
\item{\code{start}}{initial values of coefficients in IRLS algorithm}
\item{\code{K}}{penalty matrix}
}
}
\description{
This function defines the tuning parameters for the chosen loss function.
For the loss functions \code{"nils"} and \code{"med"} the input tuning parameters are preserved if the input parameter \code{tuning} is not \code{NULL}.
For \code{"kemp"} the tuning parameters are chosen as described in Kemp et al. (2012),
for \code{"mean"} they are chosen in a way that the IRLS algorithm yields the least-squares estimate.
Hence, for all four loss functions the same algorithm can be used.
}
