% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_cv.R
\name{m_cv}
\alias{m_cv}
\title{Function performing the cross-validation}
\usage{
m_cv(
  x,
  y,
  start,
  loss,
  d,
  tuning,
  maxi,
  tau,
  lambda,
  K,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{x}{for \code{\link{m}}: design matrix, for \code{m_gam()}: \code{NULL}}

\item{y}{for \code{\link{m}}: response vector, for \code{m_gam()}: \code{NULL}}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{d}{derivative of chosen loss function divided by its argument}

\item{tuning}{set of tuning parameters}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{vector of smoothing parameters which will be compared via cross-validation;
if \code{length(lambda)==2}, lambda is cross-validated on a log-scale between those values;
if \code{length(lambda>2)}, the values in \code{lambda} are cross-validated}

\item{K}{for \code{\link{m}}: quadratic penalty matrix; for \code{\link{m_gam}}: preceeding gam model with \code{fit=FALSE}}

\item{ncores}{number of cores to be used for parallelization}
}
\value{
\describe{
\item{\code{lambda_opt}}{optimal lambda chosen by cross-validation via GCV criterion}
\item{\code{cv}}{list containing the GCV scores and the estimated coefficients for the different cross-validated smoothing parameter values}
}
}
\description{
This function performs the cross-validation for models fitted with \code{\link{m}} and \code{length(lambda)>0}.
The cross-validation is done outside of the IRLS algorithm.
If \code{length(lambda)==2}, lambda is cross-validated on a log-scale between those values,
otherwise the values in \code{lambda} are cross-validated.
The GCV scores and optimal lambdas are calculated with \code{\link{m_cv_calc}}.
}
