% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_cv_calc.R
\name{m_cv_calc}
\alias{m_cv_calc}
\title{Function for calculating the GCV scores and finding the optimal lambda}
\usage{
m_cv_calc(
  x,
  y,
  start,
  loss,
  d,
  tuning,
  maxi,
  tau,
  lambda,
  K,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{x}{for \code{\link{m}}: design matrix, for \code{m_gam()}: \code{NULL}}

\item{y}{for \code{\link{m}}: response vector, for \code{m_gam()}: \code{NULL}}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{d}{derivative of chosen loss function divided by its argument}

\item{tuning}{set of tuning parameters}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{vector of possible smoothing parameters}

\item{K}{for \code{\link{m}}: quadratic penalty matrix; for \code{\link{m_gam}}: preceeding gam model with \code{fit=FALSE}}

\item{ncores}{number of cores to be used for parallelization}
}
\value{
\describe{
\item{\code{lambda_opt}}{optimal lambda chosen by cross-validation via GCV criterion}
\item{\code{lambda}}{vector of smoothing parameters which will be compared via cross-validation}
\item{\code{scores}}{GCV scores for the different values in \code{lambda}}
\item{\code{coefs}}{estimated coefficients for the different values in \code{lambda}}
}
}
\description{
This function takes a vector \code{lambda} of possible values for lambda as input and fits the according model for each lambda.
Finally, it calculates the GCV scores for each lambda from the fits and detects the optimal lambda
(meaning the lambda which yields the minimal GCV score). If the minimum is not unique, the maximal optimal lambda is chosen.
}
