% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_fit.R
\name{m_fit}
\alias{m_fit}
\title{Function fitting the model}
\usage{
m_fit(x = NULL, y = NULL, start, loss, d, tuning, maxi, tau, lambda, K)
}
\arguments{
\item{x}{for \code{\link{m}}: design matrix, for \code{m_gam()}: \code{NULL}}

\item{y}{for \code{\link{m}}: response vector, for \code{m_gam()}: \code{NULL}}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{d}{derivative of chosen loss function divided by its argument}

\item{tuning}{set of tuning parameters}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{for \code{\link{m}}: smoothing parameter, for \code{\link{m_gam}}: smoothing parameter estimation method
as required by \code{\link[mgcv]{gam}}}

\item{K}{for \code{\link{m}}: quadratic penalty matrix; for \code{\link{m_gam}}: preceeding gam model with \code{fit=FALSE}}
}
\value{
\describe{
\item{\code{coefficients}}{estimated coefficients}
\item{\code{residuals}}{residuals}
\item{\code{fitted}}{fitted values}
\item{\code{weights}}{weights of final iteration in IRLS algorithm}
\item{\code{iter}}{number of iterations}
\item{\code{j}}{internal index of tuning parameters used in last IRLS iteration}
\item{\code{converged}}{logical; if \code{TRUE}, the IRLS algorithm converged}
\item{\code{gam_object}}{for \code{\link{m}}: \code{NULL}, for \code{\link{m_gam}}: gam object in final iteration}
}
}
\description{
This function fits the model by performing the IRLS algorithm proposed by Oelker et al. (2015).
}
