% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_gam.R
\name{m_gam}
\alias{m_gam}
\title{Function performing the penalized mode, mean and median regression with lambda chosen inside the IRLS algorithm}
\usage{
m_gam(
  formula,
  data = list(),
  start = NULL,
  loss = "nils",
  tuning = NULL,
  maxi = 5500,
  tau = 10^(-5),
  B = 0,
  WB = FALSE,
  fu = function(x) x,
  type = "gumbel",
  B_fit = FALSE,
  alpha = NULL,
  method = "REML",
  ncores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{formula}{model formula as required by \code{\link[mgcv]{gam}}}

\item{data}{data set}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{tuning}{set of tuning parameters, if \code{NULL}, the tuning parameters are not set manually}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{B}{number of bootstrap samples; if \code{B=0} no bootstrap will be performed}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{B_fit}{logical; if \code{FALSE}, the estimated coefficients for each bootstrap sample are returned in \code{bs_ciraw}; if \code{TRUE}, the fitted values per sample are returned}

\item{alpha}{confidence level for bootstrap confidence intervals; if \code{alpha=NULL} the confidence level is set to 0.05}

\item{method}{smoothing parameter estimation method as required by \code{\link[mgcv]{gam}}}

\item{ncores}{number of cores to be used for parallelization}

\item{...}{further arguments which can be passed to \code{\link[mgcv]{gam}}}
}
\value{
\describe{
\item{\code{coefficients}}{estimated coefficients}
\item{\code{residuals}}{residuals}
\item{\code{fitted}}{fitted values}
\item{\code{weights}}{weights of final iteration in IRLS algorithm}
\item{\code{iter}}{number of iterations in IRLS algorithm}
\item{\code{converged}}{logical; if \code{TRUE}, the IRLS algorithm converged}
\item{\code{tuning}}{set of tuning parameters}
\item{\code{lambda}}{estimated smoothing parameters for the gam model in the last IRLS iteration}
\item{\code{gam_object}}{gam object in the last IRLS iteration}
\item{\code{bs_ci}}{bootstrap confidence intervals}
\item{\code{bs_ciraw}}{estimated coefficients or fitted values for each bootstrap sample}
\item{\code{k}}{for \code{loss="nils"}: tuning parameter in last IRLS iteration, for \code{loss="kemp"}: inverse bandwith in last IRLS iteration}
\item{\code{x}}{design matrix (generated internally in \code{\link[mgcv]{gam}})}
}
}
\description{
This function performs the penalized mode, median and mean regression where the penalization parameter lambda
is chosen inside the IRLS algorithm and the estimation of the coefficent vector and of the penalty parameter is interlaced.
}
