% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_study_m.R
\name{perform_study_m}
\alias{perform_study_m}
\title{Function performing a simulation study using \code{\link{m}}}
\usage{
perform_study_m(
  data,
  coefs,
  start = NULL,
  loss = "nils",
  tuning = NULL,
  maxi = 5500,
  tau = 10^(-5),
  lambda = 0,
  K = NULL,
  B = 0,
  B_nested = B,
  bs_ci_method = c("percentile", "rev_percentile", "et_percentile-t",
    "sym_percentile-t"),
  WB = FALSE,
  trafo_type = "identity",
  type = "gumbel",
  B_fit = FALSE,
  alpha = 0.05,
  cov = TRUE,
  exact = FALSE,
  kernel = "gaussian",
  ncores = parallel::detectCores() - 1,
  path = "./results",
  filename
)
}
\arguments{
\item{data}{list containing the data for the different repetitions of the study}

\item{coefs}{true coefficients}

\item{start}{initial values of coefficients in IRLS algorithm}

\item{loss}{type of loss function: \code{"nils"} = approach of Oelker et al. (2015),
\code{"kemp"} = approach of Kemp et al. (2012), \code{"mean"} = mean regression, \code{"med"} = approximation of median
regression}

\item{tuning}{set of tuning parameters, if \code{NULL}, the tuning parameters are not set manually}

\item{maxi}{maximal number of iterations in the IRLS algorithm}

\item{tau}{criterion for convergence of IRLS algorithm}

\item{lambda}{vector of smoothing parameters; if \code{length(lambda>1)} they will be compared via cross-validation;
if \code{length(lambda)==2}, lambda is cross-validated on a log-scale between those values;
if \code{length(lambda>2)}, the values in \code{lambda} are cross-validated}

\item{K}{quadratic penalty matrix}

\item{B}{number of bootstrap samples; if \code{B=0} no bootstrap will be performed}

\item{B_nested}{for \code{bs_ci_method = "et_percentile-t"}: number of nested bootstrap iterations per original bootstrap iteration}

\item{bs_ci_method}{vector of bootstrap confidence interval construction methods to be used: \code{"percentile"} = percentile bootstrap,
\code{"rev_percentile"} = bootstrap reverse percentile, \code{"et_percentile-t"} = equal-tailed bootstrap percentile-t,
\code{"sym_percentile-t"} = symmetric bootstrap percentile-t}

\item{WB}{logical; if \code{TRUE}, wild bootstrap instead of residual bootstrap is performed}

\item{trafo_type}{type of transformation function to be used, currently six options are implemented:
\code{"identity"}, \code{"absolute"}, \code{"sqrt_hat"}, \code{"hat"}, \code{"df"}, \code{"feng"}}

\item{type}{type of weight distribution, currently seven options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"}, \code{"mammen_two"}, \code{"webb_six"}}

\item{B_fit}{logical; if \code{FALSE}, the estimated coefficients for each bootstrap sample are returned in \code{bs_ciraw}; if \code{TRUE}, the fitted values per sample are returned}

\item{alpha}{confidence level for bootstrap confidence intervals; if \code{alpha=NULL} the confidence level is set to 0.05}

\item{cov}{logical; if \code{TRUE} and the type of loss function is either \code{"nils"} or \code{"kemp"},
the estimated asymptotic covariance matrix multiplied by k^3/n will be calculated}

\item{exact}{logical; for \code{loss="mean"}: \code{exact="TRUE"} can be chosen to calculate the exact
confidence interval under the normal distribution assumption}

\item{kernel}{for \code{loss="med"}: kernel for estimating the asymptotic covariance matrix (\code{gaussian} or \code{uniform})}

\item{ncores}{number of cores to be used for parallelization}

\item{path}{path were the .Rdata file containing the results should be saved}

\item{filename}{filename of results file}
}
\description{
This function performs a simulation study for given data, true coefficients and a specific loss function.
The results containing the true coefficients, the estimated coefficients, the fitted values, the residuals,
the asymptotic and bootstrap confidence intervals, the lengths and coverage rates of the confidence intervals,
the bandwiths used in the final IRLS iterations and the numbers of IRLS iterations are saved to an .RData file.
}
