% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_m.R
\name{results_m}
\alias{results_m}
\title{Function collecting and plotting the results of a simulation study performed with \code{\link{perform_study_m}}}
\usage{
results_m(
  results_path = "./results",
  results_filename,
  plots_path = "./plots",
  plots_filename = results_filename,
  tables_path = "./tables",
  tables_filename = results_filename
)
}
\arguments{
\item{results_path}{path to the folder which contains the results of the preceeding simulation studies
for the four different losses}

\item{results_filename}{filename (without the "_loss" ending) of the results files}

\item{plots_path}{path to the folder where the plots should be saved}

\item{plots_filename}{filename of the resulting plots}

\item{tables_path}{path to the folder where the tables should be saved}

\item{tables_filename}{filename of generated tables}
}
\description{
This function plots and saves boxplots of the estimated coefficients, of the lengths of the
asymptotic confidence intervals as well as of the bootstrap confidence intervals
and of the root mean squared errors. Furthermore, it collects and saves the coverage rates and the means
and standard deviations of the estimated coefficients.
}
