% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo_WB.R
\name{trafo_WB}
\alias{trafo_WB}
\title{Function transforming the residuals for the wild bootstrap}
\usage{
trafo_WB(
  residuals,
  fu = function(x) x,
  type = c("gumbel", "alaplace", "mammen", "normal", "rademacher", "mammen_two",
    "webb_six", "feng", "test")
)
}
\arguments{
\item{residuals}{original residuals}

\item{fu}{transformation function for the residuals}

\item{type}{type of weight distribution, currently eight options are implemented:
\code{"gumbel"}, \code{"alaplace"}, \code{"mammen"}, \code{"normal"}, \code{"rademacher"},
\code{"mammen_two"}, \code{"webb_six"}, \code{"feng"}}
}
\value{
\describe{
\item{\code{residuals_WB}}{transformed residuals}
}
}
\description{
This function transforms the residuals if \code{WB="TRUE"} is chosen.
If no transformation function for the residuals is specified, the identity function is used.
Currently, the weight distributions \code{"gumbel"} (Gumbel distribution with mode = 0, variance = 1, skewness = 1),
\code{"alaplace"} (asymmetric Laplace distribution with approx. mode = 0, variance = 1, skewness = 1),
\code{"mammen"} (continuous distribution proposed by Mammen (1993)),
\code{"normal"} (standard normal distribution), \code{"rademacher"} (Rademacher distribution),
\code{"mammen_two"} (Mammen's two-point distribution) and \code{"webb_six"} (Webb's six-point distribution) are implemented.
The parameter values are chosen such that the weight distribution has approximately
zero mode, variance one and skewness one.
}
