% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_function.R
\name{transformation_function}
\alias{transformation_function}
\title{Function returning the transformation function for wild bootstrap}
\usage{
transformation_function(
  trafo_type = c("identity", "absolute", "sqrt_hat", "hat", "df", "feng"),
  n = NULL,
  p = NULL,
  x = NULL
)
}
\arguments{
\item{trafo_type}{type of transformation function to be returned, currently six options are implemented:
\code{"identity"}, \code{"absolute"}, \code{"sqrt_hat"}, \code{"hat"}, \code{"df"}, \code{"feng"}}

\item{n}{number of observations}

\item{p}{number of coefficients (including the intercept)}

\item{x}{design matrix}
}
\value{
\describe{
\item{\code{f}}{transformation function for wild bootstrap}
}
}
\description{
This function returns the transformation function for wild bootstrap depending on the chosen type of function.
The following types are available: \code{"identity"} (identity function), \code{"absolute"} (absolute value function),
\code{"sqrt_hat"} (function deviding the residuals by the square root of 1 minus the respective diagonal entry of the hat matrix),
\code{"hat"} (function deviding the residuals by 1 minus the respective diagonal entry of the hat matrix) and
\code{"df"} (function multiplying the residuals with \code{sqrt(n/(n-p))}.
}
