% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_parallel.R
\name{try_parallel}
\alias{try_parallel}
\title{Function immitating \code{\link[base]{try}} in parallelized setting}
\usage{
try_parallel(f)
}
\arguments{
\item{f}{expression to be wrapped}
}
\description{
This function immitates the behaviour of \code{\link[base]{try}} in parallelized settings using
\code{\link[pbmcapply]{pbmclapply}} / \code{\link[pbmcapply]{pbmcmapply}}.
}
